Availability of metadata should be cached on a per-repository basis
